Add transformVector() of CGAffineTransform. Add comment for angleTo().
It's diffrent between point and vector in geometry, but now CGPoint is used to represent both point and vector. The function `func * (lhs:CGPoint, rhs:CGAffineTransform) -> CGPoint` is used for point apparently. We need a function to transform a vector, which need't the translation elements of CGAffineTransform.

I noticed the TODO comment in CGPoint+Trigonometry.swift, so I add comment for angleTo().